api/requirements/common.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
Pillow==9.4.0
psycopg2==2.9.5
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
drf-spectacular==0.25.1
uritemplate==4.1.1
jsonschema==4.17.3
inflection==0.5.1
pyrsistent==0.19.3
gunicorn==20.1.0
requests==2.28.2
charset-normalizer==3.0.1
idna==3.4
urllib3==1.26.14
certifi==2022.12.7
PyJWT==2.6.0
